+++ GPTifier/src/command_run.cpp
@@ -1,4 +1,5 @@
-#include "run_completion.hpp"
+#include "command_run.hpp"
+
 #include "params.hpp"
 #include "utils.hpp"
